+++ src/previewdef/focustree/loader.ts
@@ -55,14 +55,10 @@ export class FocusTreeLoader extends ContentLoader<FocusTreeLoaderResult> {
             .flatMap(ft => Object.values(ft.focuses))
             .flatMap(f => [...f.icon.map(i => i.icon), f.overlay])
             .value();
-        const workspaceGfxFiles = (await listFilesFromModOrHOI4('interface', { hoi4: false, recursively: true }))
-            .filter(f => f.toLocaleLowerCase().endsWith('.gfx'))
-            .map(f => 'interface/' + f);
 
         const gfxDependencies = [
             ...dependencies.filter(d => d.type === 'gfx').map(d => d.path),
             ...flatten(focusTreeDepFiles.map(f => f.result.gfxFiles)),
-            ...workspaceGfxFiles,
             ...await getGfxContainerFiles(focusGfxNames),
         ];